assignment-12/requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-tavily>=0.1.0
langgraph>=0.1.0
openai>=1.99.0
requests>=2.31.0
tiktoken>=0.5.0
PyPDF2>=3.0.0
Pillow>=10.0.0
pydantic>=2.0.0
